Updates * Fixed the input component to use the StyleInjector properly * Updated a number of form components to use the StyleInjector
