+++ repos/keg-components/src/components/form/checkbox/checkbox.js
@@ -3,6 +3,7 @@ import { Check } from 'KegIcons'
 import React, { useMemo } from 'react'
 import { CheckboxWrapper } from './checkbox.wrapper'
 import { noPropObj } from '../../../utils/helpers/noop'
+import { StyleInjector } from '@keg-hub/re-theme/styleInjector'
 
 // Styles are defined here so that they are enforced
 // Due to how the checkbox is designed, these styles should not be changed
@@ -34,6 +35,14 @@ const checkBoxStyles = {
   },
 }
 
+/**
+ * @summary A simple wrapper around the a web input element
+ * <br/> This is required because we can't just pass input to the StyleInjector
+ * <br/> It requires a React component, and input gets viewed as a variable
+ * @param {Object} props - props object. Accepts all standard <input /> props which will be passed to the input element. Additional props:
+ */
+const InputComp = props => (<input {...props} />)
+
 const Element = React.forwardRef(
   (
     { className, elProps, styles = noPropObj, icon, checked, ...props },
@@ -45,6 +54,11 @@ const Element = React.forwardRef(
         ...checkBoxStyles.icon,
       }
     }, [ checkBoxStyles, styles ])
+
+    const InjectedComp = StyleInjector(
+      InputComp,
+      { displayName: 'Checkbox', className: 'keg-checkbox' }
+    )
 
     return (
       <View
@@ -59,7 +73,7 @@ const Element = React.forwardRef(
           className='keg-checkbox-icon'
           style={checkStyle}
         /> }
-        <input
+        <InjectedComp
           className='keg-checkbox'
           {...elProps}
           {...props}
